stream messages to terminal client, rename conversation
