--- mentat/conversation.py
@@ -1,483 +1,0 @@
-import asyncio
-import logging
-from dataclasses import asdict, dataclass
-from pathlib import Path
-from textwrap import dedent
-from typing import Any, Dict, Iterable, List, Literal, Set
-
-from ipdb import set_trace
-from termcolor import cprint
-
-from .code_change import CodeChange, CodeChangeAction
-from .code_file_manager import CodeFileManager
-from .code_map import CodeMap
-from .config_manager import ConfigManager
-from .llm_api import CostTracker, check_model_availability, choose_model, count_tokens
-from .parsing import (
-    async_stream_and_parse_llm_response,
-    run_async_stream_and_parse_llm_response,
-)
-from .prompts import system_prompt
-
-logger = logging.getLogger()
-
-
-@dataclass
-class BaseMessage:
-    content: str
-
-
-@dataclass
-class Message(BaseMessage):
-    content: str
-    extra: Dict[str, Any] | None = None
-
-
-@dataclass
-class ConversationMessage(BaseMessage):
-    role: Literal["system"] | Literal["assistant"] | Literal["user"]
-    content: str
-
-
-@dataclass
-class QuestionMessage(BaseMessage):
-    content: str
-    options: Set[str] | None = None
-    default: str | None = None
-    extra: Dict[str, Any] | None = None
-
-
-@dataclass
-class AnswerMessage(BaseMessage):
-    content: str
-
-
-class MentatConversation:
-    def __init__(self, config: ConfigManager, include_code_map: bool = True):
-        self.config = config
-
-        self._main_task: asyncio.Task | None = None
-        self._stop_task: asyncio.Task | None = None
-        self._user_response_queue: asyncio.Queue[str] = asyncio.Queue()
-
-        self.messages: List[BaseMessage] = []
-        self.add_system_message(system_prompt)
-        self.code_file_manager = CodeFileManager(config=self.config)
-        self.cost_tracker = CostTracker()
-
-        # NOTE: The code map will be outdated whenever any file inside of it changes
-        if include_code_map:
-            self.code_map = CodeMap(str(self.config.git_root), token_limit=2048)
-            if self.code_map.ctags_disabled:
-                self.add_message(
-                    content=f"""
-                        There was an error with your universal ctags installation,
-                        disabling CodeMap. Reason: {self.code_map.ctags_disabled_reason}
-                    """,
-                    color="yellow",
-                )
-        else:
-            self.code_map = None
-
-        self.allow_32k = check_model_availability(config.allow_32k())
-
-        tokens = count_tokens(self.code_file_manager.get_code_message()) + count_tokens(
-            system_prompt
-        )
-        self.token_limit = 32768 if self.allow_32k else 8192
-        if tokens > self.token_limit:
-            raise KeyboardInterrupt(
-                f"Included files already exceed token limit ({tokens} /"
-                f" {self.token_limit}). Please try running again with a reduced number"
-                " of files."
-            )
-        elif tokens + 1000 > self.token_limit:
-            self.add_message(
-                f"Warning: Included files are close to token limit ({tokens} /"
-                f" {self.token_limit}), you may not be able to have a long"
-                " conversation.",
-                color="red",
-            )
-        else:
-            self.add_message(
-                f"File and prompt token count: {tokens} / {self.token_limit}",
-                color="cyan",
-            )
-
-    def add_message(self, content: str, **kwargs):
-        message = Message(content=dedent(content), extra={**kwargs})
-        self.messages.append(message)
-
-    def add_system_message(self, content: str):
-        message = ConversationMessage(role="system", content=content)
-        self.messages.append(message)
-
-    def add_user_message(self, content: str):
-        message = ConversationMessage(role="user", content=content)
-        self.messages.append(message)
-
-    def add_assistant_message(self, content: str):
-        message = ConversationMessage(role="assistant", content=content)
-        self.messages.append(message)
-
-    async def prompt_yes_no(self, content: str, default_yes: bool, **kwargs) -> bool:
-        self.messages.append(
-            Message(content=("(y/n)" if default_yes else "(y/n)"), extra=kwargs)
-        )
-        self.messages.append(
-            QuestionMessage(
-                content=content + ("(y/n)" if default_yes else "(y/n)"),
-                options={"y", "n"},
-                default="y" if default_yes else "n",
-                extra=kwargs,
-            )
-        )
-
-        while True:
-            user_input = await self._user_response_queue.get()
-            user_input = user_input.lower()
-            if user_input not in {"y", "n", ""}:
-                self.messages.append(
-                    Message(content=("(y/n)" if default_yes else "(y/n)"), extra=kwargs)
-                )
-                continue
-            break
-
-        return user_input == "y" or (user_input != "n" and default_yes)
-
-    async def get_model_response(self):
-        # why do we make a copy here?
-        _messages = self.messages.copy()
-
-        # Filter out non-conversation messages
-        messages = [m for m in _messages if isinstance(m, ConversationMessage)]
-
-        code_message = self.code_file_manager.get_code_message()
-        system_message = code_message
-
-        if self.code_map:
-            system_message_token_count = count_tokens(system_message)
-            messages_token_count = 0
-            for message in messages:
-                messages_token_count += count_tokens(message.content)
-            token_buffer = 1000
-            token_count = (
-                system_message_token_count + messages_token_count + token_buffer
-            )
-            max_tokens_for_code_map = self.token_limit - token_count
-
-            if self.code_map.token_limit:
-                code_map_message_token_limit = min(
-                    self.code_map.token_limit, max_tokens_for_code_map
-                )
-            else:
-                code_map_message_token_limit = max_tokens_for_code_map
-
-            code_map_message = self.code_map.get_message(
-                token_limit=code_map_message_token_limit
-            )
-            if code_map_message:
-                match (code_map_message.level):
-                    case "signatures":
-                        cprint_message_level = "full syntax tree"
-                    case "no_signatures":
-                        cprint_message_level = "partial syntax tree"
-                    case "filenames":
-                        cprint_message_level = "filepaths only"
-                    case _:
-                        raise Exception(
-                            f"Unknown CodeMapMessage level '{code_map_message.level}'"
-                        )
-                self.add_message(
-                    content=f"\nIncluding CodeMap ({cprint_message_level})",
-                    color="green",
-                )
-                system_message += f"\n{code_map_message}"
-            else:
-                cprint_message = [
-                    "\nExcluding CodeMap from system message.",
-                    "Reason: not enough tokens available in model context.",
-                ]
-                cprint_message = "\n".join(cprint_message)
-                self.add_message(cprint_message, color="yellow")
-
-        self.add_system_message(system_message)
-
-        model, num_prompt_tokens = choose_model(
-            [asdict(m) for m in messages], self.allow_32k
-        )
-
-        state = await async_stream_and_parse_llm_response(
-            [asdict(m) for m in messages], model, self.code_file_manager
-        )
-
-        self.cost_tracker.display_api_call_stats(
-            num_prompt_tokens,
-            count_tokens(state.message),
-            model,
-            state.time_elapsed,
-        )
-
-        self.add_assistant_message(state.message)
-        return state.explanation, state.code_changes
-
-    async def get_user_response(self):
-        user_response = await self._user_response_queue.get()
-
-        return user_response
-
-    async def get_user_filter_on_changes(self, code_changes: Iterable[CodeChange]):
-        new_changes = []
-        indices = []
-        for index, change in enumerate(code_changes, start=1):
-            # print_change(change)
-            # Allowing the user to remove rename file changes introduces a lot of edge cases
-            if change.action == CodeChangeAction.RenameFile:
-                new_changes.append(change)
-                indices.append(index)
-                self.add_message(
-                    "Cannot remove rename file change", color="light_yellow"
-                )
-                continue
-
-            self.add_message("Keep this change?", color="light_blue")
-            keep_change = await self.prompt_yes_no(
-                content="Keep this change?", default_yes=True
-            )
-            if keep_change:
-                new_changes.append(change)
-                indices.append(index)
-
-        return new_changes, indices
-
-    async def get_user_feedback_on_changes(
-        self, code_changes: Iterable[CodeChange]
-    ) -> bool:
-        self.add_message(
-            "Apply these changes? 'Y/n/i' or provide feedback.", color="light_blue"
-        )
-        user_response = await self.get_user_response()
-
-        need_user_request = True
-        match user_response.lower():
-            case "y" | "":
-                code_changes_to_apply = code_changes
-                self.add_user_message("User chose to apply all your changes.")
-            case "n":
-                code_changes_to_apply = []
-                self.add_user_message("User chose not to apply any of your changes.")
-            case "i":
-                code_changes_to_apply, indices = await self.get_user_filter_on_changes(
-                    code_changes
-                )
-                self.add_user_message(
-                    "User chose to apply"
-                    f" {len(code_changes_to_apply)}/{len(code_changes)} of your suggest"
-                    " changes. The changes they applied were:"
-                    f" {', '.join(map(str, indices))}"
-                )
-            case _:
-                need_user_request = False
-                code_changes_to_apply = []
-                self.add_user_message(
-                    "User chose not to apply any of your changes. User response:"
-                    f" {user_response}\n\nPlease adjust your previous plan and changes to"
-                    " reflect this. Respond with a full new set of changes."
-                )
-
-        if code_changes_to_apply:
-            self.code_file_manager.write_changes_to_files(code_changes_to_apply)
-            if len(code_changes_to_apply) == len(code_changes):
-                self.add_message("Changes applied.", color="light_blue")
-            else:
-                self.add_message("Selected changes applied.", color="light_blue")
-        else:
-            self.add_message("No changes applied.", color="light_blue")
-
-        if need_user_request:
-            cprint("Can I do anything else for you?", color="light_blue")
-
-        return need_user_request
-
-    ### lifecycle
-
-    @property
-    def is_stopped(self):
-        return self._main_task is None and self._stop_task is None
-
-    async def _main(self):
-        self.add_message("Type 'q' or use Ctrl-C to quit at any time.\n", color="cyan")
-        self.add_message("What can I do for you?", color="light_blue")
-
-        need_user_request = True
-        while True:
-            if need_user_request:
-                user_response = await self.get_user_response()
-                self.add_user_message(user_response)
-            explanation, code_changes = await self.get_model_response()
-
-            if code_changes:
-                need_user_request = await self.get_user_feedback_on_changes(
-                    code_changes
-                )
-            else:
-                need_user_request = True
-
-    def start(self):
-        if self._main_task:
-            logger.warning("Job already started")
-            return
-
-        async def run_main():
-            try:
-                await self._main()
-            except asyncio.CancelledError:
-                pass
-
-        def cleanup_main(_: asyncio.Task):
-            self._main_task = None
-            logger.debug("Main task stopped")
-
-        self._main_task = asyncio.create_task(run_main())
-        self._main_task.add_done_callback(cleanup_main)
-
-    def stop(self):
-        if self._stop_task is not None:
-            logger.warning("Task is already stopping")
-            return
-        if self.is_stopped:
-            logger.warning("Task is already stopped")
-            return
-
-        async def run_stop():
-            if self._main_task is None:
-                return
-            try:
-                # Wait for main task to cancel
-                self._main_task.cancel()
-                while self._main_task is not None:
-                    await asyncio.sleep(0.1)
-            except asyncio.CancelledError:
-                pass
-
-        def cleanup_stop(_: asyncio.Task):
-            self._stop_task = None
-            logger.debug("Task has stopped")
-
-        self._stop_task = asyncio.create_task(run_stop())
-        self._stop_task.add_done_callback(cleanup_stop)
-
-
-class Conversation:
-    def __init__(
-        self,
-        config: ConfigManager,
-        cost_tracker: CostTracker,
-        code_file_manager: CodeFileManager,
-        code_map: CodeMap | None = None,
-    ):
-        self.messages = []
-        self.add_system_message(system_prompt)
-        self.cost_tracker = cost_tracker
-        self.code_file_manager = code_file_manager
-        self.code_map = code_map
-        self.allow_32k = check_model_availability(config.allow_32k())
-
-        tokens = count_tokens(code_file_manager.get_code_message()) + count_tokens(
-            system_prompt
-        )
-        self.token_limit = 32768 if self.allow_32k else 8192
-        if tokens > self.token_limit:
-            raise KeyboardInterrupt(
-                f"Included files already exceed token limit ({tokens} /"
-                f" {self.token_limit}). Please try running again with a reduced number"
-                " of files."
-            )
-        elif tokens + 1000 > self.token_limit:
-            cprint(
-                f"Warning: Included files are close to token limit ({tokens} /"
-                f" {self.token_limit}), you may not be able to have a long"
-                " conversation.",
-                "red",
-            )
-        else:
-            cprint(
-                f"File and prompt token count: {tokens} / {self.token_limit}", "cyan"
-            )
-
-    def add_system_message(self, message: str):
-        self.messages.append({"role": "system", "content": message})
-
-    def add_user_message(self, message: str):
-        self.messages.append({"role": "user", "content": message})
-
-    def add_assistant_message(self, message: str):
-        self.messages.append({"role": "assistant", "content": message})
-
-    def get_model_response(self, config: ConfigManager) -> (str, list[CodeChange]):
-        messages = self.messages.copy()
-
-        code_message = self.code_file_manager.get_code_message()
-        system_message = code_message
-
-        if self.code_map:
-            system_message_token_count = count_tokens(system_message)
-            messages_token_count = 0
-            for message in messages:
-                messages_token_count += count_tokens(message["content"])
-            token_buffer = 1000
-            token_count = (
-                system_message_token_count + messages_token_count + token_buffer
-            )
-            max_tokens_for_code_map = self.token_limit - token_count
-
-            if self.code_map.token_limit:
-                code_map_message_token_limit = min(
-                    self.code_map.token_limit, max_tokens_for_code_map
-                )
-            else:
-                code_map_message_token_limit = max_tokens_for_code_map
-
-            code_map_message = self.code_map.get_message(
-                token_limit=code_map_message_token_limit
-            )
-            if code_map_message:
-                match (code_map_message.level):
-                    case "signatures":
-                        cprint_message_level = "full syntax tree"
-                    case "no_signatures":
-                        cprint_message_level = "partial syntax tree"
-                    case "filenames":
-                        cprint_message_level = "filepaths only"
-                    case _:
-                        raise Exception(
-                            f"Unknown CodeMapMessage level '{code_map_message.level}'"
-                        )
-                cprint_message = f"\nIncluding CodeMap ({cprint_message_level})"
-                cprint(cprint_message, color="green")
-                system_message += f"\n{code_map_message}"
-            else:
-                cprint_message = [
-                    "\nExcluding CodeMap from system message.",
-                    "Reason: not enough tokens available in model context.",
-                ]
-                cprint_message = "\n".join(cprint_message)
-                cprint(cprint_message, color="yellow")
-
-        messages.append({"role": "system", "content": system_message})
-
-        model, num_prompt_tokens = choose_model(messages, self.allow_32k)
-
-        state = run_async_stream_and_parse_llm_response(
-            messages, model, self.code_file_manager
-        )
-
-        self.cost_tracker.display_api_call_stats(
-            num_prompt_tokens,
-            count_tokens(state.message),
-            model,
-            state.time_elapsed,
-        )
-
-        self.add_assistant_message(state.message)
-        return state.explanation, state.code_changes
