+++ mentat/session_conversation.py
@@ -2,6 +2,8 @@ from dataclasses import dataclass
 from typing import Any, Dict, List
 from uuid import UUID, uuid4
 
+from ipdb import set_trace
+
 from .broadcast import Broadcast
